Refactoring code to move add collaborator to company view/serializer
