+++ company/tests/test_views.py
@@ -2633,3 +2633,27 @@ def test_request_identity_verification(mock_send_request_identity_message, authe
     assert response.status_code == 200
     assert mock_send_request_identity_message.call_count == 1
     assert mock_send_request_identity_message.call_args == mock.call(authed_supplier)
+
+
+@pytest.mark.django_db
+def test_add_collaborator_view(authed_client):
+    company = factories.CompanyFactory(
+        name='Test Company', date_of_creation=datetime.date(2000, 10, 10),
+    )
+    data = {
+        'sso_id': 300,
+        'name': 'Abc',
+        'company': company.number,
+        'company_email': 'abc@def.com',
+        'mobile_number': 9876543210,
+        'role': user_roles.MEMBER
+    }
+
+    url = reverse('register-company-collaborator-request')
+    response = authed_client.post(
+        url,
+        data=data
+    )
+
+    assert response.status_code == status.HTTP_201_CREATED
+    assert response.json() == {"company_email": "abc@def.com"}
